+++ local_dev_vendor.sh
@@ -1,6 +1,0 @@
-#!/usr/bin/env bash
-make vendor
-cd vendor/github.com/Keyfactor
-rm -rf keyfactor-go-client
-ln -s "$HOME/GolandProjects/keyfactor-go-client" .
-cd ../../..