+++ CHANGELOG.md
@@ -1,11 +1,22 @@
 # v2.1.6
 
+### Client
+
+#### Features
+
+* 9808312 feat(client): `keyfactor_client` now allows for global `request_timeout` to be set. Default is 30 seconds.
+
+#### Fixes
+
+* 9808312 fix(client): `keyfactor_client` now retries any 'Context Deadline Exceeded' errors.
+
 ### Certificates
 
 #### Fixes
 
 * 4d01ddd fix(certificates): `keyfactor_certificate` resources now handle enrollments requests that require approvals.
   #90
+
 
 ### Deployments